lil refactoring, too much dialogs :face_with_spiral_eyes:
