+++ apps/web/src/components/pages/settings/index.ts
@@ -1,0 +1,2 @@
+export * from "./SidebarDialog";
+export * from "./ManageDialog";
